# feat: add debug.SymbolTable into constraint system for storage efficiency of debug info