… constraint/bls12-377/r1cs_test.go
… constraint/bls12-381/r1cs_test.go
… constraint/bls24-315/r1cs_test.go
… constraint/bls24-317/r1cs_test.go
… constraint/bn254/r1cs_test.go
… constraint/bw6-633/r1cs_sparse.go
… constraint/bw6-633/r1cs_test.go
… constraint/bw6-633/solution.go
… constraint/bw6-761/r1cs.go
… constraint/bw6-761/r1cs_sparse.go
… constraint/bw6-761/r1cs_test.go
… constraint/bw6-761/solution.go
… constraint/debug_info.go
… constraint/log.go
… constraint/system.go
… constraint/tinyfield/r1cs.go
… constraint/tinyfield/r1cs_sparse.go
… constraint/tinyfield/r1cs_test.go
… constraint/tinyfield/solution.go
… debug/debug.go
… debug/symbol_table.go
… debug_test.go
… frontend/cs/r1cs/api_assertions.go
… frontend/cs/r1cs/builder.go
… frontend/cs/scs/builder.go
… internal/generator/backend/template/representations/r1cs.go.tmpl
… internal/generator/backend/template/representations/r1cs.sparse.go.tmpl
… internal/generator/backend/template/representations/solution.go.tmpl
… internal/generator/backend/template/representations/tests/r1cs.go.tmpl
… profile/profile_test.go