Clearing that from backend; don't store stack trace in string (debug info) in constraint systems, just store ids to a symbol table and format at run time.

Additionally, don't store the stack at all for `AssertIsBoolean` unless debug build tag is set since we can have a huge amount of these.

Example impact on bn254 pairing:
```
benchmark                                  old ns/op        new ns/op       delta
BenchmarkFinalExponentiationCompile-10     148012806625     42604127375     -71.22%

benchmark                                  old allocs     new allocs     delta
BenchmarkFinalExponentiationCompile-10     919164251      534771532      -41.82%

benchmark                                  old bytes        new bytes       delta
BenchmarkFinalExponentiationCompile-10     121579914912     78784978536     -35.20%
```